Move multi box overlay to own package
