+++ src/main/java/org/janelia/saalfeldlab/paintera/PainteraDefaultHandlers.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.DoubleSupplier;
 
-import bdv.fx.viewer.MultiBoxOverlayRendererFX;
+import bdv.fx.viewer.multibox.MultiBoxOverlayRendererFX;
 import bdv.fx.viewer.ViewerPanelFX;
 import bdv.viewer.Interpolation;
 import bdv.viewer.Source;
